Fix style, add test, make EllipsisMenu own component
